Made the program more modular for cleanliness reasons
Changes made:
- Broke off functions into 2 modules (image_handling.py and input_tools.py
- Removed the option to use hardcoded variables, it's not as useful now that the program is broken up
- Added lots of comments
- Fixed a visual issue printing the File Created message
